+++ pkg/host/machine_id.go
@@ -5,35 +5,12 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"runtime"
 	"strings"
-	"time"
 
 	"github.com/leptonai/gpud/pkg/file"
 	"github.com/leptonai/gpud/pkg/log"
 	"github.com/leptonai/gpud/pkg/process"
 )
-
-var currentMachineID string
-
-func init() {
-	if runtime.GOOS != "linux" {
-		return
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	var err error
-	currentMachineID, err = GetMachineID(ctx)
-	if err != nil {
-		log.Logger.Errorw("failed to get machine id", "error", err)
-	}
-}
-
-func CurrentMachineID() string {
-	return currentMachineID
-}
 
 // Returns the UUID of the machine host.
 // Returns an empty string if the UUID is not found.
@@ -47,27 +24,6 @@ func GetMachineID(ctx context.Context) (string, error) {
 		return ReadOSMachineID()
 	}
 	return uuid, nil
-}
-
-var currentDmidecodeUUID string
-
-func init() {
-	if runtime.GOOS != "linux" {
-		return
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	var err error
-	currentDmidecodeUUID, err = DmidecodeUUID(ctx)
-	if err != nil {
-		log.Logger.Errorw("failed to get UUID from dmidecode", "error", err)
-	}
-}
-
-func CurrentDmidecodeUUID() string {
-	return currentDmidecodeUUID
 }
 
 // Fetches the UUIF of the machine host, using the "dmidecode".
@@ -125,24 +81,6 @@ func extractUUID(line string) string {
 		return ""
 	}
 	return strings.TrimSpace(strings.TrimPrefix(line, "UUID: "))
-}
-
-var currentOSMachineID string
-
-func init() {
-	if runtime.GOOS != "linux" {
-		return
-	}
-
-	var err error
-	currentOSMachineID, err = ReadOSMachineID()
-	if err != nil {
-		log.Logger.Errorw("failed to get machine id", "error", err)
-	}
-}
-
-func CurrentOSMachineID() string {
-	return currentOSMachineID
 }
 
 // ref. https://github.com/google/cadvisor/blob/854445c010e0b634fcd855a20681ae986da235df/machine/info.go#L39
